+++ packages/observation-extractors/evm-observation-extractor/lib/index.ts
@@ -1,2 +1,3 @@
+export * from './BinanceRpcObservationExtractor';
 export * from './EthereumRpcObservationExtractor';
 export * from './EvmRpcObservationExtractor';