+++ packages/watcher-data-extractor/lib/extractor/EventTriggerExtractor.ts
@@ -2,19 +2,23 @@ import { DataSource } from 'typeorm';
 import * as wasm from 'ergo-lib-wasm-nodejs';
 import { blake2b } from 'blakejs';
 import { Transaction } from '@rosen-bridge/scanner';
-import { AbstractLogger, DummyLogger } from '@rosen-bridge/abstract-logger';
-import { AbstractExtractor, Block } from '@rosen-bridge/abstract-extractor';
+import { AbstractLogger } from '@rosen-bridge/abstract-logger';
+import {
+  AbstractInitializableErgoExtractor,
+  BlockInfo,
+  ErgoNetworkType,
+  OutputBox,
+  SpendInfo,
+} from '@rosen-bridge/abstract-extractor';
 
 import EventTriggerAction from '../actions/EventTriggerAction';
 import { ExtractedEventTrigger } from '../interfaces/extractedEventTrigger';
 import { JsonBI } from '../utils';
 import { EventResult } from '../types';
 
-class EventTriggerExtractor extends AbstractExtractor<Transaction> {
-  readonly logger: AbstractLogger;
+class EventTriggerExtractor extends AbstractInitializableErgoExtractor<ExtractedEventTrigger> {
   id: string;
-  private readonly dataSource: DataSource;
-  private readonly actions: EventTriggerAction;
+  protected readonly actions: EventTriggerAction;
   private readonly eventTriggerErgoTree: string;
   private readonly permitErgoTree: string;
   private readonly fraudErgoTree: string;
@@ -23,15 +27,16 @@ class EventTriggerExtractor extends AbstractExtractor<Transaction> {
   constructor(
     id: string,
     dataSource: DataSource,
+    type: ErgoNetworkType,
+    url: string,
     address: string,
     RWT: string,
     permitAddress: string,
     fraudAddress: string,
     logger?: AbstractLogger
   ) {
-    super();
+    super(type, url, address, logger);
     this.id = id;
-    this.dataSource = dataSource;
     this.eventTriggerErgoTree = wasm.Address.from_base58(address)
       .to_ergo_tree()
       .to_base16_bytes();
@@ -42,11 +47,120 @@ class EventTriggerExtractor extends AbstractExtractor<Transaction> {
       .to_ergo_tree()
       .to_base16_bytes();
     this.RWT = RWT;
-    this.logger = logger ? logger : new DummyLogger();
     this.actions = new EventTriggerAction(dataSource, this.logger);
   }
 
+  /**
+   * get Id for current extractor
+   */
   getId = () => this.id;
+
+  /**
+   * check proper data format in the box
+   *  - box ergoTree
+   *  - RWT in first token place
+   *  - widListHash in R4
+   *  - event data in R5
+   *  - widCount in R7
+   * @param box
+   * @return true if the box has the required data and false otherwise
+   */
+  hasData = (box: OutputBox): boolean => {
+    try {
+      const parsedBox = wasm.ErgoBox.from_json(JsonBI.stringify(box));
+      if (
+        box.additionalRegisters &&
+        box.additionalRegisters.R4 &&
+        box.additionalRegisters.R5 &&
+        box.additionalRegisters.R7 &&
+        parsedBox.register_value(wasm.NonMandatoryRegisterId.R4) &&
+        parsedBox.register_value(wasm.NonMandatoryRegisterId.R5) &&
+        parsedBox.register_value(wasm.NonMandatoryRegisterId.R7) &&
+        box.assets &&
+        box.assets.length > 0 &&
+        box.assets[0].tokenId === this.RWT &&
+        box.ergoTree === this.eventTriggerErgoTree
+      ) {
+        const R4Serialized = parsedBox
+          .register_value(wasm.NonMandatoryRegisterId.R4)!
+          .to_byte_array();
+        const R5Serialized = parsedBox
+          .register_value(wasm.NonMandatoryRegisterId.R5)!
+          .to_coll_coll_byte();
+        if (R4Serialized.length >= 1 && R5Serialized.length >= 12) return true;
+      }
+    } catch (e) {
+      this.logger.warn(
+        `Unexpected error occurred while checking the proper trigger data format for box ${box.boxId}: ${e}`
+      );
+    }
+    return false;
+  };
+
+  /**
+   * extract box data to proper format (not including spending information)
+   * @param box
+   * @return extracted data in proper format
+   */
+  extractBoxData = (
+    box: OutputBox
+  ): Omit<ExtractedEventTrigger, 'spendBlock' | 'spendHeight'> | undefined => {
+    try {
+      const parsedBox = wasm.ErgoBox.from_json(JsonBI.stringify(box));
+      const R4Serialized = parsedBox
+        .register_value(wasm.NonMandatoryRegisterId.R4)!
+        .to_byte_array();
+      const R5Serialized = parsedBox
+        .register_value(wasm.NonMandatoryRegisterId.R5)!
+        .to_coll_coll_byte();
+
+      const WIDsCount = parsedBox
+        .register_value(wasm.NonMandatoryRegisterId.R7)!
+        .to_i32();
+      const WIDsHash = Buffer.from(R4Serialized).toString('hex');
+      const sourceTxId = Buffer.from(R5Serialized[0]).toString();
+      const eventId = Buffer.from(blake2b(sourceTxId, undefined, 32)).toString(
+        'hex'
+      );
+      return {
+        eventId: eventId,
+        txId: box.transactionId,
+        boxId: box.boxId,
+        boxSerialized: Buffer.from(parsedBox.sigma_serialize_bytes()).toString(
+          'base64'
+        ),
+        toChain: Buffer.from(R5Serialized[2]).toString(),
+        toAddress: Buffer.from(R5Serialized[4]).toString(),
+        networkFee: BigInt(
+          '0x' + Buffer.from(R5Serialized[7]).toString('hex')
+        ).toString(10),
+        bridgeFee: BigInt(
+          '0x' + Buffer.from(R5Serialized[6]).toString('hex')
+        ).toString(10),
+        amount: BigInt(
+          '0x' + Buffer.from(R5Serialized[5]).toString('hex')
+        ).toString(10),
+        sourceChainTokenId: Buffer.from(R5Serialized[8]).toString(),
+        targetChainTokenId: Buffer.from(R5Serialized[9]).toString(),
+        sourceTxId: sourceTxId,
+        fromChain: Buffer.from(R5Serialized[1]).toString(),
+        fromAddress: Buffer.from(R5Serialized[3]).toString(),
+        sourceBlockId: Buffer.from(R5Serialized[10]).toString(),
+        WIDsCount: WIDsCount,
+        WIDsHash: WIDsHash,
+        sourceChainHeight: Number(
+          BigInt('0x' + Buffer.from(R5Serialized[11]).toString('hex')).toString(
+            10
+          )
+        ),
+      };
+    } catch (e) {
+      this.logger.warn(
+        `Unexpected error occurred while extracting trigger data for box ${box.boxId}: ${e}`
+      );
+      return undefined;
+    }
+  };
 
   /**
    * gets block id and transactions corresponding to the block and saves if they are valid rosen
@@ -54,155 +168,52 @@ class EventTriggerExtractor extends AbstractExtractor<Transaction> {
    * @param block
    * @param txs
    */
-  processTransactions = (
+  processTransactions = async (
     txs: Array<Transaction>,
-    block: Block
+    block: BlockInfo
   ): Promise<boolean> => {
-    return new Promise((resolve, reject) => {
-      try {
-        const boxes: Array<ExtractedEventTrigger> = [];
-        const txSpendIds: Array<{
-          txId: string;
-          spendBoxes: Array<string>;
-          result: string;
-          paymentTxId: string;
-        }> = [];
-        txs.forEach((transaction) => {
-          // extract event result
-          const { result, paymentTxId } = this.extractEventResult(transaction);
-          for (const output of transaction.outputs) {
-            try {
-              if (
-                output.additionalRegisters &&
-                output.assets &&
-                output.assets.length > 0 &&
-                output.assets[0].tokenId === this.RWT &&
-                output.ergoTree === this.eventTriggerErgoTree
-              ) {
-                const R4 = output.additionalRegisters.R4;
-                const R5 = output.additionalRegisters.R5;
-                const R7 = output.additionalRegisters.R7;
-                if (R4 && R5 && R7) {
-                  const outputParsed = wasm.ErgoBox.from_json(
-                    JsonBI.stringify(output)
-                  );
-                  const R4Const = outputParsed.register_value(
-                    wasm.NonMandatoryRegisterId.R4
-                  );
-                  const R5Const = outputParsed.register_value(
-                    wasm.NonMandatoryRegisterId.R5
-                  );
-                  const R7Const = outputParsed.register_value(
-                    wasm.NonMandatoryRegisterId.R7
-                  );
-                  if (R4Const && R5Const && R7Const) {
-                    const R4Serialized = R4Const.to_byte_array();
-                    const R5Serialized = R5Const.to_coll_coll_byte();
-                    if (R4Serialized.length >= 1 && R5Serialized.length >= 12) {
-                      const WIDsCount = R7Const.to_i32();
-                      const WIDsHash =
-                        Buffer.from(R4Serialized).toString('hex');
-                      const sourceTxId = Buffer.from(
-                        R5Serialized[0]
-                      ).toString();
-                      const eventId = Buffer.from(
-                        blake2b(sourceTxId, undefined, 32)
-                      ).toString('hex');
-                      boxes.push({
-                        eventId: eventId,
-                        txId: transaction.id,
-                        boxId: output.boxId,
-                        boxSerialized: Buffer.from(
-                          outputParsed.sigma_serialize_bytes()
-                        ).toString('base64'),
-                        toChain: Buffer.from(R5Serialized[2]).toString(),
-                        toAddress: Buffer.from(R5Serialized[4]).toString(),
-                        networkFee: BigInt(
-                          '0x' + Buffer.from(R5Serialized[7]).toString('hex')
-                        ).toString(10),
-                        bridgeFee: BigInt(
-                          '0x' + Buffer.from(R5Serialized[6]).toString('hex')
-                        ).toString(10),
-                        amount: BigInt(
-                          '0x' + Buffer.from(R5Serialized[5]).toString('hex')
-                        ).toString(10),
-                        sourceChainTokenId: Buffer.from(
-                          R5Serialized[8]
-                        ).toString(),
-                        targetChainTokenId: Buffer.from(
-                          R5Serialized[9]
-                        ).toString(),
-                        sourceTxId: sourceTxId,
-                        fromChain: Buffer.from(R5Serialized[1]).toString(),
-                        fromAddress: Buffer.from(R5Serialized[3]).toString(),
-                        sourceBlockId: Buffer.from(R5Serialized[10]).toString(),
-                        WIDsCount: WIDsCount,
-                        WIDsHash: WIDsHash,
-                        sourceChainHeight: Number(
-                          BigInt(
-                            '0x' + Buffer.from(R5Serialized[11]).toString('hex')
-                          ).toString(10)
-                        ),
-                      });
-                    }
-                  }
-                }
-              }
-            } catch {
-              continue;
+    try {
+      const boxes: Array<ExtractedEventTrigger> = [];
+      const spendInfoArray: Array<SpendInfo> = [];
+      txs.forEach((transaction) => {
+        // extract event result
+        const { result, paymentTxId } = this.extractEventResult(transaction);
+        for (const output of transaction.outputs) {
+          // extract output box data
+          if (this.hasData(output)) {
+            this.logger.debug(
+              `Trying to extract data from box ${output.boxId} at height ${block.height}`
+            );
+            const data = this.extractBoxData(output);
+            if (data) {
+              this.logger.debug(
+                `Extracted data ${JsonBI.stringify(data)} from box ${
+                  output.boxId
+                }`
+              );
+              boxes.push(data);
             }
           }
-          // process inputs
-          txSpendIds.push({
+        }
+        // process inputs
+        for (let index = 0; index < transaction.inputs.length; index++)
+          spendInfoArray.push({
             txId: transaction.id,
-            spendBoxes: transaction.inputs.map((box) => box.boxId),
-            result: result,
-            paymentTxId: paymentTxId,
+            boxId: transaction.inputs[index].boxId,
+            index: index,
+            extras: [result, paymentTxId],
           });
-        });
-        this.actions
-          .storeEventTriggers(boxes, block, this.getId())
-          .then(() => {
-            txSpendIds.forEach((txSpends) => {
-              this.actions
-                .spendEventTriggers(
-                  txSpends.spendBoxes,
-                  block,
-                  this.id,
-                  txSpends.txId,
-                  txSpends.result,
-                  txSpends.paymentTxId
-                )
-                .then(() => {
-                  resolve(true);
-                });
-            });
-          })
-          .catch((e) => {
-            this.logger.error(
-              `Error in storing permits of the block ${block}: ${e}`
-            );
-            reject(e);
-          });
-      } catch (e) {
-        this.logger.error(
-          `block ${block} doesn't save in the database with error: ${e}`
-        );
-        reject(e);
-      }
-    });
-  };
-
-  forkBlock = async (hash: string) => {
-    await this.actions.deleteBlock(hash, this.getId());
-  };
-
-  /**
-   * Extractor box initialization
-   * No action needed in event trigger extractors
-   */
-  initializeBoxes = async () => {
-    return;
+      });
+      if (boxes.length > 0)
+        await this.actions.insertBoxes(boxes, block, this.getId());
+      await this.actions.spendBoxes(spendInfoArray, block, this.id);
+    } catch (e) {
+      this.logger.error(
+        `Error in storing data in ${this.getId()} of the block ${block}: ${e}`
+      );
+      return false;
+    }
+    return true;
   };
 
   /**